fix: layout select
