+++ src/pages/common/Layout/layout.scss
@@ -155,6 +155,7 @@
           white-space: nowrap;
           padding: 1em;
           font-size: 14px;
+          color: #fff;
           &:hover {
             color: #fff;
           }
